chore: reorganised entities
