+++ src/auth/auth.controller.ts
@@ -20,7 +20,7 @@ import {
   ApiTags,
   ApiUnauthorizedResponse,
 } from '@nestjs/swagger';
-import { COMMON_SWAGGER_RESPONSES, AUTH_SWAGGER_RESPONSES } from 'src/swagger';
+import { COMMON_SWAGGER_RESPONSES, AUTH_SWAGGER_RESPONSES } from '../swagger';
 @Controller('auth')
 @ApiTags('auth')
 @ApiBadRequestResponse({
@@ -30,7 +30,7 @@ import { COMMON_SWAGGER_RESPONSES, AUTH_SWAGGER_RESPONSES } from 'src/swagger';
   example: COMMON_SWAGGER_RESPONSES.forbiddenResponse,
 })
 export class AuthController {
-  constructor(private authService: AuthService) {}
+  constructor(private authService: AuthService) { }
 
   @ApiCreatedResponse({ example: AUTH_SWAGGER_RESPONSES.register })
   @Public()
