+++ src/entities/cash-kick-contract.entity.ts
