+++ src/entities/cash-kick.entity.ts
