+++ src/entities/index.ts
@@ -1,0 +1,5 @@
+export * from './cash-kick-contract.entity';
+export * from './cash-kick.entity';
+export * from './payment-schedule.entity';
+export * from './user.entity';
+export * from './contract.entity';
